+++ lib/db/migrations.js
@@ -1,7 +1,7 @@
 import { ObjectID } from "bson";
 import { getInsights } from "../insights";
 import { getPostsCollection } from "../mongodb";
-import { decryptPost } from "../transformations";
+import { decryptPost } from "../security/transformations";
 import { hasCommentedOnPost } from "./reads";
 import { generateCompletion } from "./writes";
