+++ lib/db/posts/reads.js
@@ -1,6 +1,6 @@
 import { ObjectID } from "bson";
 import { getPostsCollection } from "../../mongodb";
-import { decryptPost } from "../../transformations";
+import { decryptPost } from "../../security/transformations";
 import { getUserByEmail } from "../reads";
 
 /**
