+++ lib/db/posts/writes.js
@@ -5,7 +5,7 @@ import {
   getPostsCollection,
   getStarsCollection,
 } from "../../mongodb";
-import { encryptPost } from "../../transformations";
+import { encryptPost } from "../../security/transformations";
 import { getPostById, getUserByEmail, hasAiCommentedOnPost } from "../reads";
 import { isPostOwner } from "../../validations";
 import { generateCompletion } from "../writes";
