+++ lib/security/encryption.js
