+++ lib/security/transformations.js
